separated RepositoryResult and IConnectionDetailRepository into it's own classes
