+++ src/kentaasvang.Rssh/Repositories/ConnectionDetailRepository.cs
@@ -21,17 +21,3 @@ public class ConnectionDetailRepository : IConnectionDetailRepository
     throw new NotImplementedException();
   }
 }
-
-public interface IConnectionDetailRepository
-{
-  public RepositoryResult<ConnectionDetailEntity> Insert(ConnectionDetailEntity entity);
-  public RepositoryResult<List<ConnectionDetailEntity>> GetAll();
-  public RepositoryResult<ConnectionDetailEntity> Delete(string name);
-}
-
-public class RepositoryResult<T> where T : class
-{
-  public bool Succeeded { get; set; } = false;
-  public T? Value { get; set; }
-  public string ErrorMessage { get; set; } = string.Empty;
-}
